Added a difference-printer in draw_density.
So, if you'll look at the output when you do draw_density, you can see
that at the end, (the bottom-right), there's a 14.7 unit difference
between calculated t*'s. Though it's ok for there to be a difference
away from the edges, at the edges there should be a 0 difference between
tx's of different resolutions. This is probably where we're fucking up.